pyproject.toml
[project]
name = "league-match-history-discord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "py-cord>=2.6.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ruff>=0.12.11",
]

[tool.ruff]
lint.extend-select = ["I"]